requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
pandas==2.0.3
kagglehub==0.3.13
opencv-python==4.8.1.78
matplotlib==3.8.2
scipy==1.11.4
tensorboard==2.15.1
tqdm==4.66.1
pyyaml==6.0.1
pillow==10.1.0
